fix: extract shared usePauseToggle hook and PauseToggleButton component
